requirements.txt
matplotlib==3.6.2
numpy==1.23.4
Pillow==9.3.0
Pillow==10.1.0
Requests==2.31.0
scipy==1.11.3
